Uninstaller changes made as suggested
