+++ Cargo.toml
@@ -11,4 +11,3 @@ serde_json = "1.0.128"
 sevenz-rust = { version = "0.6.1", features = ["zstd", "bzip2"] }
 zip = "2.2.0"
 zip-extract = { version = "0.2.1", features = ["bzip2", "default", "xz", "zstd", "lzma"] }
-regex = "1.5"
