+++ src/cli/subcommands.rs
@@ -1,6 +1,6 @@
-use crate::manifest::Manifest;
-use crate::scoopd::manifest_uninstall::uninstall_package;
+use crate::scoopd::manifest_uninstall;
 use crate::AlephConfig;
+use crate::{manifest::Manifest, scoopd::manifest_uninstall::remove_package_dir};
 use std::{
     fmt::Arguments,
     path::{Path, PathBuf},
@@ -29,13 +29,7 @@ impl SubCommand {
             SubCommand::Search => search_repo(config, argument),
             SubCommand::Install => install_repo_manifest(config, argument),
             SubCommand::Fetch => fetch_repo(config, argument),
-            SubCommand::Uninstall => {
-                if let Some(pkg_name) = argument {
-                    uninstall_package(config, pkg_name)
-                } else {
-                    Err("Package name required for uninstall.".to_string())
-                }
-            }
+            SubCommand::Uninstall => uninstall_package(config, argument),
             SubCommand::Rebuild => unimplemented!(""),
         }
     }
@@ -152,6 +146,32 @@ fn install_repo_manifest(config: &AlephConfig, arg: Option<&String>) -> Result<(
         manifest_installer(config, &manifest, package)?;
     }
 
+    Ok(())
+}
+
+pub fn uninstall_package(config: &AlephConfig, arg: Option<&String>) -> Result<(), String> {
+    // Ensure we have an argument.
+    let arg = arg.ok_or("Package name required for uninstall.".to_string())?;
+
+    // Split the argument string into individual package names.
+    for pkg in arg.split_whitespace().map(|s| s.trim()) {
+        // Compute the expected package directory: $HOME\Aleph\Packages\<pkg>
+        let package_path: PathBuf = config.paths.packages.join(pkg);
+        if !package_path.exists() {
+            println!("Package '{}' not found at {:?}", pkg, package_path);
+            continue; // Skip this package if the directory does not exist.
+        }
+        println!(
+            "Found package '{}' at {:?}. Proceeding with uninstall...",
+            pkg, package_path
+        );
+
+        // Call the manifest uninstallation logic.
+        // We assume that in your manifest_uninstall.rs you have a function like:
+        // `pub fn uninstall_repo_manifest(config: &AlephConfig, arg: Option<&String>) -> Result<(), String>`
+        // which handles deleting the files for a given package.
+        remove_package_dir(config, Some(&pkg.to_string()))?;
+    }
     Ok(())
 }
