+++ src/powershell/mod.rs
@@ -1,2 +1,3 @@
 pub mod installer;
+pub mod uninstaller;
 pub mod utilities;
