+++ src/scoopd/manifest_uninstall.rs
@@ -1,16 +1,21 @@
-use crate::AlephConfig;
+use crate::{powershell::uninstaller::remove_from_profile, AlephConfig};
 use std::fs;
 
-pub fn uninstall_package(config: &AlephConfig, package_name: &str) -> Result<(), String> {
+pub fn remove_package_dir(
+    config: &AlephConfig,
+    package_name: Option<&String>,
+) -> Result<(), String> {
+    // Unwrap the Option to get a &String, and then trim it to get a &str.
+    let package_name = package_name.ok_or("Package name required for uninstall.".to_string())?;
+    let package_name = package_name.trim(); // Now package_name is a &str
+
     let package_path = config.paths.packages.join(package_name);
-    if !package_path.exists() {
+    if let Ok(false) = package_path.try_exists() {
         return Err(format!(
             "Package '{}' not found at {:?}",
             package_name, package_path
         ));
     }
-
-    println!("Uninstalling package: {}...", package_name);
 
     remove_from_profile(config, package_name)?;
 
@@ -18,41 +23,5 @@ pub fn uninstall_package(config: &AlephConfig, package_name: &str) -> Result<(),
         .map_err(|e| format!("Failed to remove package directory: {}", e))?;
 
     println!("Successfully uninstalled package: {}", package_name);
-    Ok(())
-}
-
-fn remove_from_profile(config: &AlephConfig, package_name: &str) -> Result<(), String> {
-    let profile_path = config
-        .paths
-        .home
-        .join("Documents")
-        .join("PowerShell")
-        .join("Microsoft.PowerShell_profile.ps1");
-
-    let profile_content = fs::read_to_string(&profile_path)
-        .map_err(|e| format!("Failed to read PowerShell profile: {}", e))?;
-
-    println!("Original profile content:\n{}", profile_content);
-
-    let search_str = format!("\\{}\\", package_name.trim());
-    println!("Filtering out lines containing: {}", search_str);
-
-    let filtered_lines: Vec<String> = profile_content
-        .lines()
-        .filter(|line| !line.contains(&search_str))
-        .map(|line| line.to_string())
-        .collect();
-
-    let new_profile = filtered_lines.join("\n");
-
-    println!("New profile content:\n{}", new_profile);
-
-    fs::write(&profile_path, new_profile)
-        .map_err(|e| format!("Failed to update PowerShell profile: {}", e))?;
-
-    println!(
-        "Removed all lines containing \"{}\" from PowerShell profile.",
-        search_str
-    );
     Ok(())
 }
